feat(api): add dependency injection
"node-dependency-injection" library has been chosen to act as a dependency injection container.

The instances registered in the container would be the following:

* `GeniallyRepository` --> `InMemoryGeniallyRepository`
* `CreateGeniallyService` --> `CreateGeniallyService(GeniallyRepository)`
* `DeleteGeniallyService` --> `DeleteGeniallyService(GeniallyRepository)`
* `RenameGeniallyService` --> `RenameGeniallyService(GeniallyRepository)`
* `GeniallysCreateController` --> `GeniallysCreateController(CreateGeniallyService)`
* `GeniallysDeleteController` --> `GeniallysDeleteController(DeleteGeniallyService)`
* `GeniallysRenameController` --> `GeniallysRenameController(RenameGeniallyService)`

Using dependency injection helps us to be able to do tests more easily by being able to configure
the different implementations we want to use for those tests.